wpa-supplicant: import upstream patch to fix dbus API
:Release Notes:
Correct dbus API to work correctly with connman.

:Detailed Notes:
This is an backported patch from upstream which can be found at
http://lists.shmoo.com/pipermail/hostap/2013-July/028167.html and
fixes a bug within the dbus API of wpa-supplicant and causes
connman's wifi scan functionality not to work correctly as
expected. Further details about the issue can be found at
https://lists.connman.net/pipermail/connman/2013-July/014871.html

Patch was submitted to oe-core upstream project as well:
http://lists.openembedded.org/pipermail/openembedded-core/2013-July/081300.html

:Testing Performed:
* compiled tested wpa-supplicant 2.0 locally with patch applied
* verified image build on target device

:QA Notes:
Please refer to [GF-7727] about a good description how the issues
addressed with this change can be tested. Furthermore it depends on
https://gpro.palm.com/#/c/5872/

:Issues Addressed:
[GF-7727] Wifi service returns {"returnValue":true}, but without network info
[GF-8334] CCC: Wifi service returns {"returnValue":true}, but without network info

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <simon.busch@lge.com>

Change-Id: I677c81ad563c0e683f49cf6b27df0de30b5926a2
Reviewed-on: https://g2g.palm.com/2785
Tested-by: Simon Busch <simon.busch@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>