webos-image: Eliminate ntp license warning
:Release Notes:
Correct the LICENSE string for ntp to eliminate a build warning from OE.

:Detailed Notes:
The upstream recipe for ntp sets the LICENSE string to "ntp" resulting
in the following warnings being generated.

    DEBUG: Executing python function do_populate_lic
    WARNING: ntp: No generic license file exists for: ntp in any provider

The LICENSE attribute is case sensitive and the correct value is thus
"NTP".

This applies a patch from upstream (recorded in the bbappend file).

:Testing Performed:
Test build to ensure warning has gone.

:QA Notes:
No change to image occurs as a result of this change and thus no CCC
is required.

:Issues Addressed:
[GF-5521] License missing for "ntp" package

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: I152cfb5e102b77843a407354cfaef54ee99a31f2
Reviewed-on: https://g2g.palm.com/3745
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>